fix: classify import export barrel indexes
