+++ packages/oxlint-plugin-react-doctor/src/plugin/utils/resolve-relative-import-path.ts
@@ -1,0 +1,54 @@
+import fs from "node:fs";
+import path from "node:path";
+
+const MODULE_FILE_EXTENSIONS = [".ts", ".tsx", ".js", ".jsx", ".mjs", ".cjs", ".mts", ".cts"];
+
+const getExistingFilePath = (filePath: string): string | null => {
+  try {
+    return fs.statSync(filePath).isFile() ? filePath : null;
+  } catch {
+    return null;
+  }
+};
+
+const getModuleFilePathCandidates = (modulePath: string): string[] => {
+  const extension = path.extname(modulePath);
+  if (!extension) {
+    return MODULE_FILE_EXTENSIONS.map((moduleExtension) => `${modulePath}${moduleExtension}`);
+  }
+
+  const modulePathWithoutExtension = modulePath.slice(0, -extension.length);
+  if (extension === ".js") {
+    return [
+      modulePath,
+      `${modulePathWithoutExtension}.ts`,
+      `${modulePathWithoutExtension}.tsx`,
+      `${modulePathWithoutExtension}.jsx`,
+    ];
+  }
+  if (extension === ".jsx") return [modulePath, `${modulePathWithoutExtension}.tsx`];
+  if (extension === ".mjs") return [modulePath, `${modulePathWithoutExtension}.mts`];
+  if (extension === ".cjs") return [modulePath, `${modulePathWithoutExtension}.cts`];
+
+  return [modulePath];
+};
+
+const resolveModuleFilePath = (modulePath: string): string | null => {
+  const exactFilePath = getExistingFilePath(modulePath);
+  if (exactFilePath) return exactFilePath;
+
+  for (const candidateFilePath of getModuleFilePathCandidates(modulePath)) {
+    const filePath = getExistingFilePath(candidateFilePath);
+    if (filePath) return filePath;
+  }
+
+  return null;
+};
+
+export const resolveRelativeImportPath = (filename: string, source: string): string | null => {
+  const importPath = path.resolve(path.dirname(filename), source);
+  const directFilePath = resolveModuleFilePath(importPath);
+  if (directFilePath) return directFilePath;
+
+  return resolveModuleFilePath(path.join(importPath, "index"));
+};
