+++ packages/react-doctor/tests/run-oxlint/bundle-size.test.ts
@@ -112,5 +112,22 @@ describe("runOxlint", () => {
         true,
       );
     });
+
+    it("flags import-then-export index barrels", async () => {
+      const projectDir = setupReactProject(tempRoot, "import-export-barrel-index-module", {
+        files: {
+          "src/components/Button.tsx": "export const Button = () => null;\n",
+          "src/components/index.ts": "import { Button } from './Button';\n\nexport { Button };\n",
+          "src/import-directory.tsx": "import { Button } from './components';\nvoid Button;\n",
+        },
+      });
+
+      const hits = await collectRuleHits(projectDir, "no-barrel-import");
+
+      expect(hits).toHaveLength(1);
+      expect(hits[0]?.filePath.replaceAll("\\", "/").endsWith("src/import-directory.tsx")).toBe(
+        true,
+      );
+    });
   });
 });
